kafka_app/requirements.txt
Flask==1.1.1
pytest==5.4.0
pytest-asyncio==0.14.0
pytest-mock==3.2.0
pytest-trio==0.6.0
pytest-tornasync==0.6.0
kafka-python==2.0.1
